requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Image Processing
Pillow==10.3.0

# Numerical Computing
numpy==1.26.4

# WSGI Server for Production
gunicorn==21.2.0
Flask-Cors

# gRPC
grpcio==1.64.1
protobuf==5.27.1